pyproject.toml
[tool.poetry]
name = "grpc-whisper-client"
version = "0.1.0"
description = "Client for grpc server at https://github.com/agent-whisper/grpc-whisper"
authors = ["Fariz Tumbuan <fariz.tumbuan@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
pyaudio = "^0.2.13"
loguru = "^0.7.0"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
